Break out process related extensions into ProcessExtensions
